Update patch set 13
Patch Set 13:

> > I don't understand this change:
 > This changes tries to work on the premise that if the application
 > is not debuggable and if either the caller or callee do not have
 > catch then the callee context is done on exception throw.
 > 
 > > 1) References to try/catch that optimizing does not handle yet.
 > This is only to explicitly capture the premise on which this works.
 > You are right, that catches are not allowed by optimizing in either
 > caller or callee. When that support is eventually added, we still
 > want this code to work and hence why the checks for catch blocks
 > are explicitly added in inliner. Another approach is to simply
 > assert it but I found it cleaner implementing it this way.

I'm not a big fan of putting infrastructure based on something we don't know how is going to be implemented. You can still implement the logic you want to implement without it.

 > 
 > > 2) How can that work with stack traces?
 > It does not work with stack traces actually and hence why I
 > explicitly check that graph is not debuggable. If we want to make
 > it work with stack traces, we can update runtime to better
 > determine location that is throwing by taking a look not only at
 > dex offset, but also at inline info to recover stack.

If you cannot have an accurate stack, you cannot do what this CL is doing. The debuggable flag is orthogonal to  the stack trace of exceptions, which are created regardless. If Quick does not preserve the stack trace, please upload a test that shows the problem.

The inline info is indeed what we should use, but is unused yet. I'm working on it.

Patch-set: 13
Reviewer: Nicolas Geoffray <1038443@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0