Update patch set 3
Patch Set 3:

Rebased.  

I am still baffled by the exact sequence needed to remove the compare and cond, and replace the If with a CompareIf.

If I replace the If with a CompareIf, then calling RemoveInstruction on the cond and the compare either dies with 
art/compiler/optimizing/nodes.cc:503] Check failed: instruction->GetEnvUses().IsEmpty() when removing the compare after I removed the cond.

If I remove the compare first, it dies with art/compiler/optimizing/nodes.cc:502] Check failed: instruction->GetUses().IsEmpty() while removing the compare.

What am I missing here?  It makes sense to remove the Cond first, as that will remove the use of the Compare.  What needs to be done with the env_uses on the compare?

Patch-set: 3
Reviewer: Mark P Mendell <1036869@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0