Update patch set 2
Patch Set 2:

Thank you for the review. I do not have a reproducer for long-to-double. I'm trying to find it during weekend but I'm not sure that I will be able to do that. I guess that for flow the incorrect conversion happens due to double rounding, first time when we sum high and low part we have a rounding and then when we round double to float. These two rounding may come to incorrect result. For double we do not have the second conversion so I think it might work correctly. I updated long-to-double to have the same implementation. If you'd like i can revert long-to-double back.

Patch-set: 2
Label: Verified=0