Update patch set 1
Patch Set 1:

Some issues that I remember encountering with the linker so far:
* It expects long headers for eh_frame for 64-bit.  This is not necessary and just wastes space.
* Line iteration was broken. (fixed now)
* Inline strings are not supported.
* ULEB and variable length data also can not be stored in debug_info.
* It breaks if ULEB is not using minimal encoding.
* The data can be nested only one level deep.
* I believe multiple compilation units also do not work.
* All tags are expected to have low_pc and high_pc attributes.

This is not meant to be a criticism of the existing code.  I am generally very happy with code which is implementing the precise minimum of features needed with the rest left to be implemented in future as needed.  However, since it appears I am going to be the one to implement the rest in the future, I feel I should justify the path I want to take.

Patch-set: 1