Update patch set 3
Patch Set 3: Code-Review+1

Update according to review. I failed to find a bad long value and I doubt I'll find it as I think there is no problem with long-to-double. So the question is still: do you want me to roll-back changes for long-to-double or we want to keep the similar code for float and double?

Patch-set: 3
Label: Code-Review=+1