Update patch set 1
Patch Set 1:

> While using tls is obviously working, I'm not sure I fully
 > understand why you need this state. You say that the debugger is
 > listenting to DexPcMoved only (therefore not breakpoints?), why
 > would you notify the method entry?

A debugger can request to receive all METHOD_ENTRY events for a class. It can also ask to receive BREAKPOINT at pc 0x0 of a method in this class.
Therefore, we must report both events (METHOD_ENTRY and BREAKPOINT) at the same time, in the same JDWP packet.

In our JDWP implementation, we do the following:
- listen to kMethodEnter instrumentation event to report METHOD_ENTRY
- listen to kDexPcMoved instrumentation event to report BREAKPOINT & SINGLE_STEP.

Because the interpreter will post instrumentation events kMethodEnter then kDexPcMoved, we need that state to not report the BREAKPOINT event twice.

Patch-set: 1