Update patch set 1
Patch Set 1:

Sent via email, then I realized I should put this here.

I tracked down why my x86 shift changes failed.  It is due to the use of Location::Any() for the LHS of a shift operation.  In the baseline compiler (caused by –debuggable), the LHS is on the stack, and so we generate for this code:
shl-int v0, v1, v4
      0x00001251:               D3642414               shl [esp + 20], cl
      0x00001255:               FF742414               push [esp + 20]
      0x00001259:               8F442410               pop [esp + 16]
Note that we clobber V1 (ESP+20) before we assign to V0 (ESP+16).

The code seems to code from:
      0 1 i37 LoadLocal [ v1 ]<|@
      0 1 i38 LoadLocal [ v4 ]<|@
      0 1 i39 Shl [ i37 i38 ]<|@
      0 0 v40 StoreLocal [ v0 i39 ]<|@

My code is:
    case Primitive::kPrimInt: {
      locations->SetInAt(0, Location::Any());
      // The shift count needs to be in CL or a constant.
      locations->SetInAt(1, Location::ByteRegisterOrConstant(ECX, op->InputAt(1)));
      locations->SetOut(Location::SameAsFirstInput());

My assumption is that someone would ensure the value is properly handled.

Do I:
•	Let you fix this bug?
•	Change the CodeGenerator to remember the is_baseline flag and set the Location to be Register() for baseline?
•	Change the OC to never use baseline, but use Quick for debuggable?
•	<Some other choice>?

Patch-set: 1