Update patch set 2
Patch Set 2:

> What do you mean by too large/noisy? Are you seeing the buffer fill
 > up at too high a rate? Or is it just too visually noisy?
 > 
 > It's entirely normal for a trace to have several things the user
 > isn't immediately interested in. The extra data in those components
 > is often useful for domain experts analyzing a trace.

Compare the following traces I collected:
1. Using a separate ATRACE_TAG_BOOT:
    https://x20web.corp.google.com/~ruhler/boottrace/40000_tag.html
2. Moving boot tracing to use tags DALVIKVM and ALWAYS as suggested, with DALVIKVM, ALWAYS, and ACTIVITY_MANAGER tags enabled:
    https://x20web.corp.google.com/~ruhler/boottrace/4041_tag.html

In (1) it is easy to see boot time is taken by Zygote and starting System Services. In (2), it isn't hard to see Zygote because it is near the top, but you have to search to see the System Services part, and if I didn't know that was something important to look for, I could easily miss it in the noise.

Also, it looks like boot in (2) takes a couple of seconds longer than (1). Could this be due to the additional tracing tags enabled? Maybe it's unrelated.

Patch-set: 2
Reviewer: Richard Uhler <1057373@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0