Update patch set 1
Patch Set 1:

This affect the code addresses which are written in dwarf data.

It feels like the right thing, but I am not completely sure as we use elf32 for 64-bit architectures.  I am not quite sure what that means.  Does the process also run in 32-bit address space?  Can we relocate beyond 4GB?

In any case, this is what DWARF says about address: Represented as an object of appropriate size to hold an address on the target machine
(DW_FORM_addr). The size is encoded in the compilation unit header (see Section 7.5.1).
This address is relocatable in a relocatable object file and is relocated in an executable file or
shared object.

Since the address size is encoded in headers, at least the tools will know how much data to read so there should be no major misinterpretation of data either way.

Patch-set: 1
Label: Verified=0