Update patch set 1
Patch Set 1:

(1 comment)

Note that this approach to supporting relative dex dependencies assumes all the dex dependencies reside in a common directory.

I talked with Andreas about this change, but I'm still unclear on why this should be checked in if it isn't being used and will want cleanup when it starts being used anyway. Am I missing something? Is the goal to start generating oat files with dex dependency info ASAP, so that when the time comes to do something with that info it will exist in the wild?

Patch-set: 1
Reviewer: Richard Uhler <1057373@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0