Update patch set 5
Patch Set 5:

Oh!  I forgot that the segments might need to be aligned.  I somehow never even managed get elf file from dex2oat so I could not run readelf on it.

I have initially created extra segments because it seemed easier.  In PS2 I have tried to at least move them towards the LOAD_R segment since that seems like the right thing to do, and I was worried odd things might be happening at the end of elf file or if there is gap in the mapped space.

After some more experimentation today I have discovered we in fact need the extra segment (without some major changes).  The code is somewhat arbitrarily split into two parts - Init() and Write(), and the debug sections are only available in the later.  Therefore when LOAD_R is being setup, we do not know anything about eh_frame.  So we need to create extra segment in Write() when we do have it.

I find the split to Init and Write odd, but the reason for it seems to be that eh_frame includes addresses of functions - which are addresses to .text.  Therefore the part of Elf file up to .text must be setup before eh_frame can be even created.  I would prefer to solve this dependency differently, but this CL probably is not the time or place for it.

The other reason I created extra segments was that I assumed that all the segments must be non-overlapping.  PT_GNU_EH_FRAME is probably supposed to be overlapping with PT_LOAD as you found from Clang & GCC.

Patch-set: 5
Label: Verified=0