Update patch set 1
Patch Set 1:

I mentioned this to David the other day.
Currently I think this can only be an issue for the CFI information, which may become inaccurate. If nothing 'looks' at the code before the recorded PC we should be fine otherwise, but it is simply safer to explicitly mark those sections. This type of issue can be a pain to debug in a JIT ;)

Patch-set: 1
Label: Verified=0