Update patch set 1
Patch Set 1:

> > Does this work correctly wrt/ inlining?
 > 
 > I don't see why there would be any problems with inlining.
 > 
 > 32-bit and 64-bit boot.oat on Nexus 9 are reduced by 28KiB and 52
 > KiB, respectively.

What's at the bottom of the stack frame of the caller when we inline? I believe that question wasn't resolved, yet. I believe it's working right now because optimizing currently cancels inlining when the dex cache might be needed and it's not the same as of the "caller." But in general you're making a change here now that we discussed before and postponed (else you *should* have dusted off my WIP).

Patch-set: 1
Label: Verified=0