Update patch set 2
Patch Set 2: Code-Review-1

I'll -1 though. It's possible to completely remove the parameter, and I think it's worth a discussion. It's useless to pass the BootClassLoader.getInstance in. No class will have that as its class-loader. For the boot classpath, the field will be null. So as long as we agree that this only ever checks for something that's not BCP, then a null-check is sufficient.

Patch-set: 2
Reviewer: Andreas Gampe <1041833@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1