Update patch set 1
Patch Set 1:

Two copies of similar code is better than three :-)  I will have no hard feelings if this gets killed in favor of the last competitor.  Actually, I would not be surprised of more copies of this existed in the codebase.

They still differ in how they achieve the goal (hard-coded path vs find).  Doing find over the repository seems nasty, but does the job.  I think the best solution would be glob, but that apparently is not an option on Android.

My eagerness to test this for all architectures seemed misplaced, as the build server only checks out prebuilts for arm.  Oh well :-)

Patch-set: 1