Update patch set 2
Patch Set 2:

Do you mean compiler flag?  Is that necessary?

What should the bug say?

I am happy just always not mentioning the x86 floating point registers.

Quote from the DWARF spec: "The default rule for all columns before interpretation of the initial instructions is the undefined rule. However, an ABI authoring body or a compilation system authoring body may specify an alternate default value for any or all columns."

In other words, even without any initial opcodes in the CIE, the debuggers should have reasonable defaults which should follow the architecture's ABI.

For the integer registers, I think it is important not to rely on any defaults and enumerate the initial state explicitly because we tend not to match the standard ABIs.

For the x86 floating point registers, the Dwarf default is undefined, the ABI's default will be undefined (floating points are never callee saved in x86, right?) and ART wants them to be undefined.  So I am not too worried.

Patch-set: 2