Update patch set 12
Patch Set 12:

This patch addresses the concern that we either build for all ISAs or none. To specify the list of ISAs to be built on the host side, modify ART_COMPILER_HOST_SUPPORTED_ARCH in art/build/Android.common.mk. To specify the list of ISAs to be built for the target, modify ART_COMPILER_TARGET_SUPPORTED_ARCH

Patch-set: 12
Reviewer: Aravind Subramanian <1067278@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0