Update patch set 1
Patch Set 1:

Still running test combinations to figure out what to turn off for 138-2, as I'm using no-dex-file-fallback to ensure we are *not* falling back when we shouldn't.

Also, the old structural check is superseded by this. I'm not sure whether I should rip it (and the now failing test 131) out or leave it, as a fall-back for when actual class-path analysis will take place.

But PTAL.

Patch-set: 1
Label: Verified=0