Update patch set 1
Patch Set 1:

Wait, you completely remove the capability of setting values with debugger: how does it not break related JDWP tests ??? Can you explain why you need to do that?

The alternative way is to deoptimize the stack to set values in a frame. However the support is not done yet. We should have dropped this code only after that step.

Patch-set: 1