Update patch set 1
Patch Set 1:

(4 comments)

LGTM.

There might be other cases where we are too agressive:

int b;
for (int i = 0; i < n; i++) {
  if (i < length) b = a[i];
  ...
}

Test is probably contrived, but we would still add deoptimizations, because the branch is not an early exit. Should we also check if the 'if' is based on the induction variable?

Patch-set: 1
Reviewer: Nicolas Geoffray <1038443@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0