Update patch set 2
Patch Set 2:

I didn't know that Mir2Lir::HandleEasyFloatingPointDiv did this in Quick.  Any idea why that is ARM only?  Care to fix Quick as well to be for all architectures.

I am not sure that the CanDivideByReciprocalMultiply{Float,Double} are right.  The code that I put in also checked the range of the exponent.

Patch-set: 2
Reviewer: Mark P Mendell <1036869@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0