Update patch set 2
Patch Set 2:

I will follow up.  Might be a day or two. Unfortunately, I am not an expert in IEEE FP.
The bad case may be denorm_1 / denorm_1, which should return 1.0, but if we replace with a multiply, there would be a loss of precision.

I will play with this to confirm and add some new test cases.

Patch-set: 2