Update patch set 1
Patch Set 1:

> Why not properly revert my CL?

Your addition of DisableGVNDependentOptimizations() to ApplyGlobalValueNumberingGate() seems logical. It's not strictly necessary but it felt a bit of a hack to rely on "temp_.gvn.gvn == nullptr" in EliminateDeadCodeGate(). That check is really intended for the case when GVN is disabled in the pass driver but DCE isn't (an odd situation to start with).

Patch-set: 1