Update patch set 2
Patch Set 2: Code-Review+1

Q: Any oat-file-assistant change?
A: I would suggest to not change oat_file_assistant, because it needs to extract gzipped odex file to read its content. But that is needed only when odex is not present and dex2oat is triggered. This dex2oat command is changed to patchoat by installd if the *.odex.gz file exists. In this case dex2oat could not succeed in case the dex file is stripped (contains no class files). In other words the gzipped odex support is needed only for precompiled odex files and it would make the oat assistant even more comlicated.

Q: Shouldn't we get rid of ElfFile's need of a file descriptor to allow in-memory decompression and patching oat files?
A: It does not seem easy to fix. We can make it later if needed as a separate patch. We should keep in mind that a deflated oat file might take more RAM if it is not backed by a file.

Patch-set: 2
Reviewer: Yevgeny Y Rouban <1037563@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1