Update patch set 1
Patch Set 1:

(1 comment)

The problem is in the existing code, not in the test. For example, art_quick_lock_object should load GP, but it doesn't. It should be done by the ENTRY macro, but the problem is that it uses .cpload $t9 for that which is wrong. .cpload is for Mips32. On Mips64 .cpsetup should be used. I will upload another patch which fix this.

Patch-set: 1
Label: Verified=0