Update patch set 1
Patch Set 1:

> (8 comments)
 > 
 > My main comment is we should consider visiting inlined methods by
 > default in the StackVisitor.

I didn't make it a default because I wanted to catch all StackVisitor implementations (ie would lead to compilation error if you don't pass the flag).

Also it's unclear that the default prevails. The main user of StackVisitor is the GC and it doesn't care about the inline frames.

 > Ignoring them should be a special case
 > and well documented in the code.
 > 
 > What do you plan to visit inlined methods by the way? Will it be
 > encoded in the oat file or inferred from the stack? Can we read
 > vregs from these methods?

It will be encoded in the oat file (the CodeInfo metadata associated with each method compiled with optimizing). And yes, we can read vregs from these methods.

Patch-set: 1