Update patch set 1
Patch Set 1:

(4 comments)

Looks pretty good.  EAX needs to be a temp for the slow path because it is where the Method* is currently passed.
      codegen->GenerateStaticOrDirectCall(invoke_->AsInvokeStaticOrDirect(), EAX);

So either EAX needs to be a temp, or it needs to be allocated as an input.  On the other hand, the call will clobber EAX anyways.

Patch-set: 1