Update patch set 1
Patch Set 1:

Given how the crash log works, I guess we do have to take this. But please make this configurable (boolean input...).

I guess for the bug it is OK to merge the outputs, as we might not need to know exactly which large object it doesn't map to, but in general that may not be the case.

Patch-set: 1
Label: Code-Review=0