Update patch set 7
Patch Set 7:

The net benefit depends on the heap layout just before collection. Our internnal benchmark shows that the average sweep time is reduced by 4.36% while the max sweep time is decreased by 15.63% while the average pause time is decreased by 5.89%, and the max pause time is reduced by 23.93%. If the heap is distributed sparsely or the layout is very long, then the benefit is obvious. If the heap is compact, the improvement will be minor accordingly. The reason of benefit for pause time is that the collection is faster, then the heap layout will be more compact after a while, in other words, the footprint will be reduced, which will make sure the scanning dirty cards in pause time more efficient. The remaining two opportunities are parallel sweep which we are doing now and the TODO item described in the comments. Specifically, for runs of objects which are 256B, 512B, 1K, and 2K, we can walk the bitmap more efficiently, since the object size is multiple of bitmap words. This benefit may be small, because the runs of large objects are fewer. Thanks.

Patch-set: 7