Update patch set 9
Patch Set 9: Code-Review-1

(Just to make sure, after Roland's +2).

Let's chat on how we should implement try/catch in the long term before/if we submit this. The change is pretty invasive in the compiler and might not be in line with a proper/full design.

Patch-set: 9
Reviewer: Nicolas Geoffray <1038443@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1