Update patch set 6
Patch Set 6:

That's a nit trick :-)

Before getting this in, I think we need an overall strategy for arch specific optimization. ARM also uploaded a CL that introduces such backend specific nodes.

I'd keep this CL around, to submit it in time when we have put a proper infrastructure.

Patch-set: 6
Label: Verified=0