Update patch set 1
Patch Set 1: Code-Review-1

Changing the bootclasspath for the core image is not a trivial change. Why did we not remove it already if it is not used?

I believe there must be a good reason why it's in the boot image, I just can't recall why.

Adding Brian and Andreas to get feedback on this change.

Patch-set: 1
Reviewer: Sebastien Hertz <1029223@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1