Update patch set 2
Patch Set 2:

> It looks like you should be able to refactor FillCalleeSaves into a
 > default implementation to avoid duplication (except for x86).

I could also just factor out the loop which is the same for core and fp regs with the notable exception of x86 fp regs. That will allow me to keep the extra DCHECK() for thumb2 and the extra bits on x86-64. Wondering if a method with a loop can be ALWAYS_INLINE...

Patch-set: 2