Update patch set 4
Patch Set 4:

> > Uploaded patch set 4.
 > 
 > This uses a kCfg... flag in runtime/globals.h. I think the code is
 > clean enough that we don't need to use `dlopen()`. What do you
 > think?
 > 
 > Also, I restored support for disassembly on target when not in
 > production mode. We think it will be useful. Any arguments against
 > that?
 > 

That's actually pretty neat. Now that I'm looking at it, I'm just thinking how to get rid of the DISASSEMBLY macro entirely. What we could do is piggy-back on kIsDebugBuild and disable the entire visualizer on release builds, including the disassembler. But this is worth a wider discussion.

 > Ideally, I'd like to take some time a little later to try to
 > clean-up argument passing for the printing mechanisms.

Do you think you could also have a look at the way the instruction addresses are collected? There's really no reason why the codegen should know anything about the visualizer. OptimizingCompiler could create the Instruction->range map, pass it to Codegen::Compile to have filled out (if not null) and then run the visualizer inside PassInfo with that info. Wouldn't that be cleaner?

Patch-set: 4
Label: Verified=0