Update patch set 4
Patch Set 4:

> > > Uploaded patch set 4.
 > >
 > > This uses a kCfg... flag in runtime/globals.h. I think the code
 > is
 > > clean enough that we don't need to use `dlopen()`. What do you
 > > think?
 > >
 > > Also, I restored support for disassembly on target when not in
 > > production mode. We think it will be useful. Any arguments
 > against
 > > that?
 > >
 > 
 > That's actually pretty neat. Now that I'm looking at it, I'm just
 > thinking how to get rid of the DISASSEMBLY macro entirely. What we
 > could do is piggy-back on kIsDebugBuild and disable the entire
 > visualizer on release builds, including the disassembler. But this
 > is worth a wider discussion.

I'm open to suggestions for simplifications. I quickly tried a `dlopen()` solution. I'm sure it is doable but it requires a little digging for details that I am not familiar with (Path of the lib on host/target? Is there a different `dlopen()` with bionic?).

 > 
 > > Ideally, I'd like to take some time a little later to try to
 > > clean-up argument passing for the printing mechanisms.
 > 
 > Do you think you could also have a look at the way the instruction
 > addresses are collected? There's really no reason why the codegen
 > should know anything about the visualizer. OptimizingCompiler could
 > create the Instruction->range map, pass it to Codegen::Compile to
 > have filled out (if not null) and then run the visualizer inside
 > PassInfo with that info. Wouldn't that be cleaner?

Yes. I had something like that in mind.
The (little) chicken and egg problem right now is that the HGraphVisualizer needs a CodeGenerator& in its constructor.
Ideally I'd like to have one top-level structure created by the OptimizingCompiler that includes configuration of the debug output (output, visualizer_enabled, cfg filename, etc.), and have this passed down to users.

I wanted to defer that to another patch because I wanted to avoid modifying other objects/logic in this patch. But if you prefer I can do that in this patch.

Patch-set: 4