Update patch set 1
Patch Set 1:

> To clarify (for anyone else) based on our hangout chat:
 > 
 > --include-debug-symbols means "include all available debugging
 > information" (this is unchanged by this CL; note that
 > --include-debug-symbols implies --include-cfi)
 > 
 > --include-cfi means the essential parts (like .text),
 > unwinding(.debug_frame), and function names(.symtab and .strtab). 
 > The last part is what this CL changes.  Function names where not
 > included before, but they will be now.
 > 
 > The intention is that --include-cfi should provide good mid-ground
 > between no debug information and all debug information.  To be
 > precise, it should include the information necessary to at least
 > produce a decent backtrace.

To complete the information from the discussion. Originally it was the reverse (gen-gdb-info implied include-debug-symbols; I missed David changing this... :-) ). The middle ground is debatable (previously it was having symtab and debug_line info, so being able to attribute a PC, but unable to unwind, which is enough for a basic perf run or getting top of Java stack). But in the end we both agree that to most sensible thing is likely to generate *all* debug info and let a stripping tool take out what you don't want.

Patch-set: 1
Reviewer: Andreas Gampe <1041833@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0