Update patch set 2
Patch Set 2:

How can it be safe to skip around parallel moves? We might miss some that should be executed before entering the loop header.

Also, it bugs me a bit that we need to now handle it in the code generators. At this point, isn't it cleaner/easier to just add a block before the loop header?


I think you can easily add an extra block before the loop header, that contains all the deopt and the new compare instruction.

Patch-set: 2