Update patch set 2
Patch Set 2: Code-Review+1

Not super happy with it as it violates the contract of the interface trampoline.

Curiously, the comments in the assembly files indicate that we used to push the dex method index before. Could you go through the git history and figure out what introduced the change?

Patch-set: 2
Reviewer: Andreas Gampe <1041833@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1