Update patch set 1
Patch Set 1: -Verified

> What's the effect on host compile time. That was the reason it's
 > off at the moment. It took 3 min last time we timed it.

ah. buzbee claimed that clang 3.4 couldn't compile this, not that it couldn't compile it in a reasonable amount of time. you build this a bunch of times, the fastest of which (aarch64) takes about 1.5mins but the slowest of which (thumb?) takes about 3mins.

did anyone raise a clang bug for this?

Patch-set: 1
Reviewer: Elliott Hughes <1003224@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0