Update patch set 1
Patch Set 1: -Code-Review

> > What's the effect on host compile time. That was the reason it's
 > > off at the moment. It took 3 min last time we timed it.
 > 
 > ah. buzbee claimed that clang 3.4 couldn't compile this, not that
 > it couldn't compile it in a reasonable amount of time. you build
 > this a bunch of times, the fastest of which (aarch64) takes about
 > 1.5mins but the slowest of which (thumb?) takes about 3mins.
 > 
 > did anyone raise a clang bug for this?

I think the state was that it may have compiled with Clang 3.4, but nobody ever was patient enough. The last try was end of last year, and it was the host that we were most interested in (because we can't even compile with GCC anymore, and we need Clang for annotalysis).

I don't remember if we have a bug for this. Stephen knows about it, and we wanted to give it to Linaro to figure out, but I think nobody got around to actually do it.

Patch-set: 1
Reviewer: Andreas Gampe <1041833@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0