Update patch set 1
Patch Set 1:

> Mips64 boots with PS2.
 > 
 > @Andreas: There are couple of other places where mips32 diverges in
 > handling the callee-saves. I am not sure whether there is more
 > working to be done:
 > 
 > 1. art_quick_invoke_stub - arm saves GP regs, while mips does not

This is not necessary. Quick treats them as callee-saves, so will spill and restore them by itself. As vmarko mentioned, this is more an inefficiency in the arm backend.

 > 2. jni compiler - there is a pending patch for saving GP registers
 > on mips, but perhaps we should save FP regs, too.

I don't think so. The native code should correctly spill and restore, so when we get back to the transition the fp registers should be back in place. The compiler uses the normal exception delivery path, which will now spill the fp regs to the stack and make them visible to delivery.

 > 
 > Also, as mips GCC may use FP callee-saves for integer values, when
 > it runs out of GP regs (as in this case it did in ExecuteGoToImpl<false,false>),
 > perhaps there are other paths that we need to consider.

I don't think so (but convince me otherwise...). Managed code doesn't touch the registers, so only transitions are interesting. In "normal" transitions, the native code should spill/restore correctly (or it's not following the ABI).

The issue here was that ART aggressively violated the ABI in native code by clobbering (for error detection).

Patch-set: 1
Label: Verified=0