Update patch set 2
Patch Set 2:

> Could you explain this a bit more, please? I don't see why (wrong)
 > inlining should be an issue, as you should never be able to get to
 > actually executing the code. Either the receiver is null, and there
 > should be an NPE, or the conversion should fail beforehand.

Wrong inlining is fine, I agree. The problem here is how we lookup the method, which is based on the vtable index (for virtual calls) of the resolved method in the dex file, and not on the signature. So the helper can crash if we don't do things properly.

On the other hand, inlining increases the generated code size, so we should probably avoid doing it when we know it will fail.

 > 
 > (To me it looks like DCE should take care of this, though I guess
 > you want to avoid generating the CFG for it just to remove it?)

Yeah, we should add some DCE/folding support for handling this.

Patch-set: 2