Update patch set 10
Patch Set 10:

(14 comments)

Nice work. Sorry for the late review. There's a lot of code in this change that is not technically related to lambdas. It'd be nice to isolate them into different changes, so that they don't get affected by potential reverts.

Also I'd appreciate if you hold off submitting the change. I haven't done a full review, being kind of distracted by changes in the CL not related to lambdas.

Patch-set: 10
Label: Verified=0