Update patch set 1
Patch Set 1:

My thinking is:  The current java compiler is producing just a filename without any path.  That makes it difficult to find the source code in any debugger.  Android Studio has the project open so it has a list of all files involved, so it can sort of do that mapping.  But it seems nasty to me.  I can envision a future where we might want to modify the java compiler to output full file path.  And which point I would not want dex2oat to neither produce garbage nor forcefully crash.  It is just "what if" to make the code a bit more robust.  If you are not happy with it, I can just abandon it.  It does not serve any purpose other than trying to be a bit more robust.

Regarding testing - we would have to modify the java compiler to do just this and then see if it works.

Patch-set: 1