Update patch set 11
Patch Set 11:

> I rebased this CL on the top of the tree.  dex2oat fails to compile
 > /system/app/BasicDreams/BasicDreams.apk because of failures in the
 > register allocator:

I can fix this failure in dex2oat by actually using the is_baseline parameter in CodeGeneratorMIPS64::SetupBlockedRegisters(), but we still seem to have clobbering of MIPS64 S- (saved) registers somewhere else, so, the generated code isn't functional with this "fix".

If I understand the latest changes correctly, the failure is triggered by the new AddTemp(visitor->GetMethodLocation()) calls in CodeGenerator::CreateCommonInvokeLocationSummary(), which increase the pressure on the available registers. But maybe that understanding is incorrect. I wonder how we manage to run out of so many MIPS64 V-, A- and T- registers (2+8+4=14) that aren't blocked in CodeGeneratorMIPS64::SetupBlockedRegisters() and need to additionally grab an S- register, whereas there are many fewer registers available on x86 and x86 is somehow OK with that. All this logic in the register allocator and the parallel mover and such should efficiently spill registers onto stack and resolve the register pressure, right?

I'm probably missing something here.
Any help?

Patch-set: 11