Update patch set 2
Patch Set 2:

It sounds like we should both increase this warning flag instead of removing it, and increase the guard page size. ASan knows how to triage segmentation faults and can report "stack-overflow" instead of a generic SEGV, but it still needs a large enough guard page to catch it.

Patch-set: 2