Update patch set 1
Patch Set 1:

> Regression test?

I'm trying to create one, but it seems this code path (setting an int/reference field from an immediate on x86-64) is never taken by the compiler, at least during the compilation of the core image or in ART tests.

For an unknown reason, it seems a parallel move is inserted between a HConstant instruction (in the entry basic block) and the H{Instance, Static}FieldSet instruction (in a later basic block).

Patch-set: 1
Label: Verified=0