Update patch set 2
Patch Set 2: Code-Review+2

Not super happy with the name. There is no "runtime throw failure," or at least it doesn't match this. What you tag is "will throw NPE at runtime at an aget-object" - which is in and of itself not a "failure."

Patch-set: 2
Reviewer: Andreas Gampe <1041833@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2