Update patch set 16
Patch Set 16:

> Examples of needing inputs that aren't standard inputs:
 > - HNeg fp wants both the fp input and also the constant for an xor
 > mask.  We want to have HNegX86 fp, HLCFCP(0x800...).  We also need
 > a temp location to load the value loaded from the constant area,
 > but that is trivial.
 > - HPackedSwitch needs have the switch value, as well as the base of
 > the constant area (HX86PackedSwitch sw_value HComputeBaseMethodAddress)
 > - min/max/abs also wants constants from the constant area.

OK, maybe those can be treated separately, and are the only problematic ones? For the xor mask, surely having another X86 specific instruction will work.

 > 
 > The idea was that adding this backend input is transparent to the
 > register allocator.

Yeah, but it won't be transparent for other phases that could expect an HAdd to only have two operands. It would actually be a nice property to not have this backend specific phase be last. Things like GVN will want to see it.

 > 
 > Invokes and returns are actually requirements on ParallelMoves as
 > noted.  I am certainly not qualified to add the needed hints to the
 > RA as to how HLCFCP operands can be properly loaded to the right
 > place. 

Then don't worry about it. It looks to me that whatever you'll try to implement in the backend will be equivalent to the hints in RA.

> What could work is having EmitMove generate a move from a
 > HLCFCP instruction (rather than the constant itself), but the
 > parallel move generator wouldn't know how to address the base of
 > the constant area without some external help.
 > 
 > I don't want to generate any code when visiting a HCLFCP node, as
 > that will involve loading into a register, which is probably not
 > what I want to generate.

Sure, you can do something similar to NeedsMaterialization.

> I want to be able to detect uses of the
 > constant area while generating code for (certain) other
 > instructions.  The classic example of this is HAdd, which wants to
 > add directly from the constant area.

Patch-set: 16