Update patch set 2
Patch Set 2:

> > Why?!?! Why does LON think it's a good idea to persist the usage
 > of
 > > that horrible flag? Nobody runs with this flag by default.
 > >
 > > Why were the tests not "duplicated," to run Quick & Optimizing?
 > For
 > > an example, see the JNI compiler test, which thanks to Nicolas
 > > tests the compiler *and* Generic JNI.
 > 
 > Alright, I'll duplicate the tests to run with both Quick &
 > Optimizing. :)

Just to make it clear: I just think it's a lot better to run them in parallel. 1) We don't have a continuous test setup with the flag on the Android build bots AFAIK. 2) This forces everyone to run them (and because of the oatdump test we do have the time, anyways... :-)). I'm just supportive of Optimizing here... :-)

Patch-set: 2