Update patch set 1
Patch Set 1:

(1 comment)

SANITIZE_TARGET is used by first doing a normal build, and then a SANITIZE_TARGET build on top of that (without clobbering). This will keep an uninstrumented dex2oat from the first build.

Sounds crazy, but it's the best we could come up with in the limitations of the build system.

Patch-set: 1
Label: Verified=0