Update patch set 8
Patch Set 8:

> (5 comments)
 > 
 > Just a few comments here and there.
 > 
 > One suggestion though - why not add the arch prefix only to -START?
 > It seems unnecessarily verbose to write it on every single line.

I started by doing that, but felt the changes were too invasive for a first step.
From a quick look again, doing that correctly would require the fnProcessLine functions to be updated to return info about the arch, and the decision on whether to ignore the arch to be done in `SplitStream`.
Now that I think about it again, I like that solution because the decision of ignoring a line is not taken in the parsing code. This will be required if we want to add more smartness, as described in a previous comment.
The downside is that the code `file_format/c1visualizer/parser.py` would need to be extended as well with something it never uses.

I'm ok to rework that if you have a preference. I would suggest sticking with the current solution involving less code for now, and updating this as needed. I suspect we will learn more about use-cases and requirements as we start implementing arch-specific tests.

 > 
 > And a nit: how about making it all uppercase? CHECK-START-ARM64,
 > for example?

Fine with me. Done.

Patch-set: 8