Update patch set 2
Patch Set 2:

> When the new macro is undefined, we are exactly in the state we
 > were before the asan annotations commit.

But I don't think that's a good state. I think it's better to have a nested #if and define empty macros when asan is active but ART is told not to sanitize.

Have you tested that Valgrind works on asan-ed builds? Otherwise there's potential runtime breakage that we could easily avoid here.

Patch-set: 2
Label: Verified=0