Update patch set 3
Patch Set 3:

> (1 comment)
 > 
 > I do not see how the notion of a "backward branch" affects any of
 > this. You should be able to reorder the code to change all
 > "forward" branches to "backward" and vice versa (adding gotos as
 > needed) with no change in behavior.

The tested constraint (B7) talks about a double execution of the same new-instance, and registers with the result of those executions. The only way to encode this is a backwards branch.

The tests for constraint B6 (which is the more generic one) would test the generic forward ones.

The test *is* important. Verifiers are conservative, approximating algorithms. So something that is perfectly fine at runtime with a loop that has multiple uninitialized objects in-flight may have to be approximated in a way that rejects that code. Now the constraints are written in words that are technically describing the runtime. As such, a constraint like B7, implicitly talking about backwards branches, is required to match the verifier with the spec (otherwise people could complain that our verifier rejects legal dex files). And if we have such a constraint, it should be tested.

Patch-set: 3
Label: Verified=0