Update patch set 1
Patch Set 1:

> > > > This won't stop building the libraries in 64-bit, right?
 > > >
 > > > No.
 > > > But the 64-bit libraries will be built only if you run mm/mmm,
 > or
 > > > required by other modules.
 > > > HOST_PREFER_32_BIT behaves the same way.
 > >
 > > Normally mm is our preferred way of operation, though.
 > >
 > > But it's the Mac, so I don't care too much...
 > It's fine to build both 32-bit and 64-bit libraries on Mac, even
 > though only the 32-bit ones are used if you run the executables.
 > Probably this is preferred, as it ensures 64-bit still builds and
 > later on you can easily switch back to 64-bit if it's allowed...

We won't switch in the foreseeable future, as our whole memory management of Java heaps would have to change. Not in the next couple of years.

So you have build time overhead (debug and non-debug) that is totally unnecessary.


But as stated: it's the mac, I don't care, go ahead and merge it!

Patch-set: 1