Update patch set 11
Patch Set 11:

I see why you added the arch to each line and not just the START but how about this...

In _extractLine, try to parse all the CHECK-START(-ARCH) combinations and store the arch that matched with the TestCase object. Then you just need to change the RunTest logic to skip TestCases which don't match the target arch specified on the command line and you'll have avoided the parser issue.

Patch-set: 11
Label: Verified=0